+++ basketserver/src/main/java/org/cryptimeleon/incentivesystem/basketserver/exceptions/BasketNotFoundException.java
@@ -1,7 +1,0 @@
-package org.cryptimeleon.incentivesystem.basketserver.exceptions;
-
-public class BasketNotFoundException extends BasketServiceException {
-    public BasketNotFoundException() {
-        super("Basket not Found");
-    }
-}