+++ services/basket/src/main/java/org/cryptimeleon/incentive/basket/BasketApplication.java
@@ -1,0 +1,13 @@
+package org.cryptimeleon.incentive.basket;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BasketApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BasketApplication.class, args);
+    }
+
+}