+++ services/issue/src/test/java/org/cryptimeleon/incentive/services/issue/IssueApplicationTests.java
@@ -1,0 +1,13 @@
+package org.cryptimeleon.incentive.services.issue;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class IssueApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}